cosmos_vector/requirements.txt
azure-cosmos>=4.7.0
azure-identity>=1.15.0
openai>=1.0.0
